analysis/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-surprise==1.1.3
scipy==1.10.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.2
zipp==3.18.1
